screens: nuke youtube music
Current home screen now supports search filters which would act the same
as this now dead screen if you enable "music_songs" filter.

This screen will now be the feed for your subscriptions, soon.
